added new class ShowsExamples.cs (extracted tests from ApiExamples.cs)
